dari panji 2/12/24 4.40

+++ database/migrations/2024_12_01_085228_create_projects_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->unsignedBigInteger('client_id');
             $table->string('title');
             $table->text('description');
-            $table->decimal('budget');
+            $table->double('budget');
             $table->dateTime('deadline');
             $table->enum('status',['open','in_progres','cancelled'])->default('open');
             $table->binary('detail');
